renames files to camel casing
