+++ extension/src/background/messageListener/handlers/addAccount.ts
