+++ extension/src/background/messageListener/handlers/addAssetsList.ts
