+++ extension/src/background/messageListener/handlers/cacheAssetDomain.ts
