+++ extension/src/background/messageListener/handlers/cacheAssetIcon.ts
