+++ extension/src/background/messageListener/handlers/changeAssetVisibility.ts
