+++ extension/src/background/messageListener/handlers/confirmMigratedMnemonicPhrase.ts
