+++ extension/src/background/messageListener/handlers/confirmMnemonicPhrase.ts
