+++ extension/src/background/messageListener/handlers/confirmPassword.ts
