+++ extension/src/background/messageListener/handlers/createAccount.ts
