+++ extension/src/background/messageListener/handlers/getCachedAssetIcons.ts
