+++ extension/src/background/messageListener/handlers/getHiddenAssets.ts
