+++ extension/src/background/messageListener/handlers/getMemoRequiredAccounts.ts
