+++ extension/src/background/messageListener/handlers/getMigratedMnemonicPhrase.ts
