+++ extension/src/background/messageListener/handlers/getMnemonicPhrase.ts
