+++ extension/src/background/messageListener/handlers/grantAccess.ts
