+++ extension/src/background/messageListener/handlers/handleSignedHwPayload.ts
