+++ extension/src/background/messageListener/handlers/importHardwareWallet.ts
