+++ extension/src/background/messageListener/handlers/loadAccount.ts
