+++ extension/src/background/messageListener/handlers/loadLastAccountUsed.ts
