+++ extension/src/background/messageListener/handlers/loadRecentAddresses.ts
