+++ extension/src/background/messageListener/handlers/loadSettings.ts
