+++ extension/src/background/messageListener/handlers/makeAccountActive.ts
