+++ extension/src/background/messageListener/handlers/modifyAssetsList.ts
