+++ extension/src/background/messageListener/handlers/recoverAccount.ts
