+++ extension/src/background/messageListener/handlers/rejectAccess.ts
