+++ extension/src/background/messageListener/handlers/rejectTransaction.ts
