+++ extension/src/background/messageListener/handlers/removeCustomNetwork.ts
