+++ extension/src/background/messageListener/handlers/removeTokenId.ts
