+++ extension/src/background/messageListener/handlers/resetExperimentalData.ts
