+++ extension/src/background/messageListener/handlers/showBackupPhrase.ts
