+++ extension/src/background/messageListener/handlers/signAuthEntry.ts
