+++ extension/src/background/messageListener/handlers/signFreighterTransaction.ts
