+++ extension/src/background/messageListener/handlers/signOut.ts
