+++ extension/src/background/messageListener/handlers/signTransaction.ts
