+++ extension/src/background/messageListener/handlers/updateAccountName.ts
