+++ extension/src/background/messageListener/popupMessageListener.ts
@@ -11,58 +11,58 @@ import { SessionTimer } from "background/helpers/session";
 import { publicKeySelector } from "background/ducks/session";
 import { EntryToSign, MessageToSign, TokenToAdd } from "helpers/urls";
 
-import { fundAccount } from "./handlers/fund-account";
-import { createAccount } from "./handlers/create-account";
-import { addAccount } from "./handlers/add-account";
-import { importAccount } from "./handlers/import-account";
-import { importHardwareWallet } from "./handlers/import-hardware-wallet";
-import { makeAccountActive } from "./handlers/make-account-active";
-import { updateAccountName } from "./handlers/update-account-name";
-import { addCustomNetwork } from "./handlers/add-custom-network";
-import { removeCustomNetwork } from "./handlers/remove-custom-network";
-import { editCustomNetwork } from "./handlers/edit-custom-network";
-import { changeNetwork } from "./handlers/change-network";
-import { loadAccount } from "./handlers/load-account";
-import { getMnemonicPhrase } from "./handlers/get-mnemonic-phrase";
-import { confirmMnemonicPhrase } from "./handlers/confirm-mnemonic-phrase";
-import { confirmMigratedMnemonicPhrase } from "./handlers/confirm-migrated-mnemonic-phrase";
-import { recoverAccount } from "./handlers/recover-account";
-import { showBackupPhrase } from "./handlers/show-backup-phrase";
-import { confirmPassword } from "./handlers/confirm-password";
-import { grantAccess } from "./handlers/grant-access";
-import { rejectAccess } from "./handlers/reject-access";
-import { handleSignedHwPayload } from "./handlers/handle-signed-hw-payload";
-import { addToken } from "./handlers/add-token";
-import { signTransaction } from "./handlers/sign-transaction";
-import { signBlob } from "./handlers/sign-blob";
-import { signAuthEntry } from "./handlers/sign-auth-entry";
-import { rejectTransaction } from "./handlers/reject-transaction";
-import { signFreighterTransaction } from "./handlers/sign-freighter-transaction";
-import { addRecentAddress } from "./handlers/add-recent-address";
-import { loadRecentAddresses } from "./handlers/load-recent-addresses";
-import { loadLastUsedAccount } from "./handlers/load-last-account-used";
-import { signOut } from "./handlers/sign-out";
-import { saveAllowList } from "./handlers/save-allow-list";
-import { saveSettings } from "./handlers/save-settings";
-import { saveExperimentalFeatures } from "./handlers/save-experimental-features";
-import { loadSettings } from "./handlers/load-settings";
-import { getCachedAssetIcon } from "./handlers/get-cached-asset-icons";
-import { cacheAssetIcon } from "./handlers/cache-asset-icon";
-import { getCachedAssetDomain } from "./handlers/get-cached-domain";
-import { cacheAssetDomain } from "./handlers/cache-asset-domain";
-import { getMemoRequiredAccounts } from "./handlers/get-memo-required-accounts";
-import { resetExperimentalData } from "./handlers/reset-experimental-data";
-import { addTokenId } from "./handlers/add-token-id";
-import { getTokenIds } from "./handlers/get-token-ids";
-import { removeTokenId } from "./handlers/remove-token-id";
-import { getMigratableAccounts } from "./handlers/get-migratable-accounts";
-import { getMigratedMnemonicPhrase } from "./handlers/get-migrated-mnemonic-phrase";
-import { migrateAccounts } from "./handlers/migrate-accounts";
-import { addAssetsList } from "./handlers/add-assets-list";
-import { modifyAssetsList } from "./handlers/modify-assets-list";
-import { getIsAccountMismatch } from "./handlers/get-is-account-mismatch";
-import { changeAssetVisibility } from "./handlers/change-asset-visibility";
-import { getHiddenAssets } from "./handlers/get-hidden-assets";
+import { fundAccount } from "./handlers/fundAccount";
+import { createAccount } from "./handlers/createAccount";
+import { addAccount } from "./handlers/addAccount";
+import { importAccount } from "./handlers/importAccount";
+import { importHardwareWallet } from "./handlers/importHardwareWallet";
+import { makeAccountActive } from "./handlers/makeAccountActive";
+import { updateAccountName } from "./handlers/updateAccountName";
+import { addCustomNetwork } from "./handlers/addCustomNetwork";
+import { removeCustomNetwork } from "./handlers/removeCustomNetwork";
+import { editCustomNetwork } from "./handlers/editCustomNetwork";
+import { changeNetwork } from "./handlers/changeNetwork";
+import { loadAccount } from "./handlers/loadAccount";
+import { getMnemonicPhrase } from "./handlers/getMnemonicPhrase";
+import { confirmMnemonicPhrase } from "./handlers/confirmMnemonicPhrase";
+import { confirmMigratedMnemonicPhrase } from "./handlers/confirmMigratedMnemonicPhrase";
+import { recoverAccount } from "./handlers/recoverAccount";
+import { showBackupPhrase } from "./handlers/showBackupPhrase";
+import { confirmPassword } from "./handlers/confirmPassword";
+import { grantAccess } from "./handlers/grantAccess";
+import { rejectAccess } from "./handlers/rejectAccess";
+import { handleSignedHwPayload } from "./handlers/handleSignedHwPayload";
+import { addToken } from "./handlers/addToken";
+import { signTransaction } from "./handlers/signTransaction";
+import { signBlob } from "./handlers/signBlob";
+import { signAuthEntry } from "./handlers/signAuthEntry";
+import { rejectTransaction } from "./handlers/rejectTransaction";
+import { signFreighterTransaction } from "./handlers/signFreighterTransaction";
+import { addRecentAddress } from "./handlers/addRecentAddress";
+import { loadRecentAddresses } from "./handlers/loadRecentAddresses";
+import { loadLastUsedAccount } from "./handlers/loadLastAccountUsed";
+import { signOut } from "./handlers/signOut";
+import { saveAllowList } from "./handlers/saveAllowList";
+import { saveSettings } from "./handlers/saveSettings";
+import { saveExperimentalFeatures } from "./handlers/saveExperimentalFeatures";
+import { loadSettings } from "./handlers/loadSettings";
+import { getCachedAssetIcon } from "./handlers/getCachedAssetIcons";
+import { cacheAssetIcon } from "./handlers/cacheAssetIcon";
+import { getCachedAssetDomain } from "./handlers/getCachedDomain";
+import { cacheAssetDomain } from "./handlers/cacheAssetDomain";
+import { getMemoRequiredAccounts } from "./handlers/getMemoRequiredAccounts";
+import { resetExperimentalData } from "./handlers/resetExperimentalData";
+import { addTokenId } from "./handlers/addTokenId";
+import { getTokenIds } from "./handlers/getTokenIds";
+import { removeTokenId } from "./handlers/removeTokenId";
+import { getMigratableAccounts } from "./handlers/getMigratableAccounts";
+import { getMigratedMnemonicPhrase } from "./handlers/getMigratedMnemonicPhrase";
+import { migrateAccounts } from "./handlers/migrateAccounts";
+import { addAssetsList } from "./handlers/addAssetsList";
+import { modifyAssetsList } from "./handlers/modifyAssetsList";
+import { getIsAccountMismatch } from "./handlers/getIsAccountMismatch";
+import { changeAssetVisibility } from "./handlers/changeAssetVisibility";
+import { getHiddenAssets } from "./handlers/getHiddenAssets";
 
 const numOfPublicKeysToCheck = 5;
